remove check for patching
Currently the check for xccdf_org.ssgproject.content_rule_security_patches_up_to_date
requires that it reaches out for remote resources, which will not be
available in a disconnected environment. The check will also fail for a
number of other reasons, which can be valid host configurations, and
requires manual intervention to discover per: https://access.redhat.com/solutions/3899241

Given we already have visibility of a host patch state, this commit
disables this check until we can handle it in a sane way.